+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.1b4 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- ``SearchableText`` index for News Article content type now includes object keywords (fixes `brasil.gov.portal#155 <https://github.com/plonegovbr/brasil.gov.portal/issues/155>`_).
+  [hvelarde]
 
 
 2.1b3 (2017-07-05)